renovate the plugin
instead including the upload functionailty in an iframe directly include
it in the form. The file input has no name attribute so it will not be
up uploaded when submitting.
When clicking the button the file will by uploaded asynchronously and a
notification  bar is shown to the user with the result.

script and css is included via an adjunct is plain vanilla javascript.